CS:PP: Week 4: Programming Assignment: World Map and minor fixes

+++ Computer Science Programming with a Purpose/hello/Week 4 Programming Assignments/ShannonEntropy.java
@@ -2,8 +2,8 @@ public class ShannonEntropy {
     public static void main(String[] args) {
         // Read from 1 to m integers from StdIn
         int m = Integer.parseInt(args[0]);
-        int counter = 0;
-        double[] p = new double[m + 1];
+        double counter = 0.0;
+        int[] p = new int[m + 1];
 
         // Read integers from StdIn and fill in tracking array
         while (!StdIn.isEmpty()) {
@@ -17,7 +17,7 @@ public class ShannonEntropy {
         // Calculate and print Shannon Entropy
         double SE = 0.0;
         for (int i = 0; i < m + 1; i++) {
-            if (p[i] != 0.0) {
+            if (p[i] != 0) {
                 double pi = p[i] / counter;
                 SE += (-pi * (Math.log(pi) / Math.log(2)));
             }
